Use Header type for arguments
